+++ components/topContainer/topContainer.wxss
@@ -1,50 +1,57 @@
 .top-container {
-	z-index: 9999;
+  z-index: 9999;
 }
 
 .absolute {
-	position: absolute;
+  position: absolute;
 }
 
 
 .top-content {
-	display: flex;
-	flex-direction: row;
-	align-items: center;
-	padding-left: 20px;
-	position: relative;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  padding-left: 20px;
+  position: relative;
 }
 
 .top-back-icon {
-	width: 28rpx;
-	height: 28rpx;
-	border: 4rpx solid #333333;
-	transform: rotate(45deg) scale(.7);
-	border-top: none;
+  width: 28rpx;
+  height: 28rpx;
+  border: 4rpx solid #333333;
+  transform: rotate(45deg) scale(.7);
+  border-top: none;
   border-right: none;
 }
 
 .top-title {
+  font-family: 'JinNanJunJunTi-Bold';
+  font-size: 34rpx;
+  line-height: 34rpx;
   display: inline-block;
   text-align: center;
-	position: absolute;
-	right: 50%;
-	transform: translateX(50%);
+  position: absolute;
+  right: 50%;
+  transform: translateX(50%);
 }
-.top-title-text{
+
+.top-title-text {
   display: inline-block;
   position: relative;
   /* left: 50%; */
   /* transform: translateX(-50%); */
 
 }
-.white{
+
+.white {
   color: white;
 }
-.borderWhite{
+
+.borderWhite {
   border-color: white;
 }
-.fixed{
+
+.fixed {
   position: fixed;
   z-index: 9999;
 }
